+++ web_app/src-capacitor/android/app/src/main/java/io/github/bentalese/dashydora/MainActivity.java
@@ -1,0 +1,5 @@
+package io.github.bentalese.dashydora;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}